use replacePlaceholdersInJson for findEndDates
